% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_predictors.R
\name{add_growing_predictors}
\alias{add_growing_predictors}
\title{Function to add growing predictors}
\usage{
add_growing_predictors(
  time_frame = NULL,
  variables = NULL,
  category = NULL,
  stats = c(mean = mean, min = min, max = max),
  output_file = TRUE,
  log_file = TRUE,
  check_size_only = FALSE,
  last_chunk_completed = NULL
)
}
\arguments{
\item{time_frame}{A time frame object}

\item{variables}{A vector or list of variable names}

\item{category}{A vector or list of category names}

\item{stats}{A list of summary stats}

\item{output_file}{Logical, whether or not to output to file or console}

\item{log_file}{Logical, whether or not to write to log}

\item{last_chunk_completed}{Integer, the previous chunk completed if chunking is used}
}
\description{
Function to add growing predictors
}
