% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_outcomes.R
\name{add_rolling_outcomes}
\alias{add_rolling_outcomes}
\title{Add rolling outcomes}
\usage{
add_rolling_outcomes(
  time_frame = NULL,
  variables = NULL,
  category = NULL,
  lookahead = lubridate::hours(48),
  window = lookahead,
  stats = c(mean = mean, min = min, max = max),
  impute = FALSE,
  output_file = TRUE,
  log_file = TRUE,
  check_size_only = FALSE,
  last_chunk_completed = NULL
)
}
\arguments{
\item{time_frame}{A time frame object}

\item{variables}{A vector or list of variable names}

\item{category}{A vector or list of category names}

\item{lookahead}{A lubridate instant, how far into the future to predict outcome}

\item{window}{A lubridate instant, break/interval to divide lookahead}

\item{stats}{A list of summary stats}

\item{impute}{Carry forward last value}

\item{output_file}{Logical, whether or not to output to file or console}

\item{log_file}{Logical, whether or not to write to log}

\item{check_size_only}{Logical, returns expected number of rows for predictor or predictor chunk if applicable}

\item{last_chunk_completed}{Integer, the previous chunk completed if chunking is used}
}
\description{
Add rolling outcomes
}
