% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine_output}
\alias{combine_output}
\title{New combine_output function}
\usage{
combine_output(
  time_frame,
  ...,
  files = NULL,
  include_files = TRUE,
  use_output_folder = TRUE,
  dplyr_join = dplyr::inner_join,
  log_file = TRUE
)
}
\description{
New combine_output function
}
