% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpmodels_internal.R
\name{gpm_add_predictors_internal}
\alias{gpm_add_predictors_internal}
\title{New furrr-enabled add_predictors function
Internal only}
\usage{
gpm_add_predictors_internal(
  time_frame = NULL,
  variables = NULL,
  category = NULL,
  lookback = lubridate::hours(48),
  window = lookback,
  stats = c(mean = mean, min = min, max = max),
  impute = TRUE,
  output_file = TRUE,
  log_file = TRUE,
  check_size_only = FALSE,
  filename_prefix = "",
  ...
)
}
\description{
New furrr-enabled add_predictors function
Internal only
}
