% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpmodels_internal.R
\name{gpm_calc}
\alias{gpm_calc}
\title{New internal helper function}
\usage{
gpm_calc(
  groups,
  temporal_id,
  temporal_variable,
  temporal_value,
  temporal_time,
  lookback_converted,
  dots,
  window_converted,
  fixed_data_of_interest,
  temporal_data_of_interest,
  stats,
  impute,
  pb,
  all_temporal_vars,
  missing_value_frame,
  strategy
)
}
\description{
New internal helper function
}
