% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpmodels_internal.R
\name{gpm_define_steps}
\alias{gpm_define_steps}
\title{New internal helper function}
\usage{
gpm_define_steps(
  groups,
  temporal_id,
  step,
  step_units,
  max_length,
  baseline,
  interval,
  max_step_times_per_id,
  lookback_converted,
  window_converted,
  output_folder,
  log_file
)
}
\description{
New internal helper function
}
