% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpmodels.R
\name{pre_dummy_code}
\alias{pre_dummy_code}
\title{Function that converts categorical temporal predictors into dummy variables}
\usage{
pre_dummy_code(
  time_frame = NULL,
  numeric_threshold = 0.5,
  variables = NULL,
  save_time_frame = TRUE
)
}
\description{
Note that you can you can use this to dummy code variables with numerical values
where the values are supposed to map to categorical levels (e.g, 1 means high and 2
means low).
}
\details{
Either provide a threshold (defaults to 0.5) or provide a vector of variables.
If you supply a vector of variables, this takes precedence over the numeric threshold.
}
